# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/student/faculty": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "get faculties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/student/habit": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "get habits",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HabitDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/student/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "get student profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "update student detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "student update",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateStudentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    }
                }
            }
        },
        "/student/{id}/habit": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "update student habit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "habit update",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateHabitDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    }
                }
            }
        },
        "/student/{id}/picture": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "upload profile picture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "profile picture",
                        "name": "profile_picture",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "upload multiple pet pictures,test this out in postman",
                        "name": "pet_pictures",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/student/{id}/preference": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "update student preference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "preference update",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePreferenceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.HabitDTO": {
            "type": "object",
            "properties": {
                "pet_habit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PetHabit"
                    }
                },
                "room_care_habit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoomCareHabit"
                    }
                },
                "sleep_habit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SleepHabit"
                    }
                },
                "smoke_habit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SmokeHabit"
                    }
                },
                "study_habit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StudyHabit"
                    }
                }
            }
        },
...
        "dto.UpdateHabitDTO": {
            "type": "object",
            "required": [
                "pet_habit_id",
                "room_care_habit_id",
                "sleep_habit_id",
                "smoke_habit_id",
                "study_habit_id"
            ],
            "properties": {
                "pet_habit_id": {
                    "type": "string"
                },
                "room_care_habit_id": {
                    "type": "string"
                },
                "sleep_habit_id": {
                    "type": "string"
                },
                "smoke_habit_id": {
                    "type": "string"
                },
                "study_habit_id": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePreferenceDTO": {
            "type": "object",
            "properties": {
                "other_preference": {
                    "type": "string"
                },
                "preferred_gender": {
                    "type": "string"
                },
                "preferred_pet_habit_id": {
                    "type": "string"
                },
                "preferred_room_care_habit_id": {
                    "type": "string"
                },
                "preferred_sleep_habit_id": {
                    "type": "string"
                },
                "preferred_smoke_habit_id": {
                    "type": "string"
                },
                "preferred_study_habit_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateStudentDTO": {
            "type": "object",
            "required": [
                "faculty",
                "firstname",
                "lastname"
            ],
            "properties": {
                "biography": {
                    "type": "string"
                },
                "facebook_url": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "twitter_url": {
                    "type": "string"
                }
            }
        },
...
        "model.PersonalHabit": {
            "type": "object",
            "properties": {
                "pet_habit": {
                    "$ref": "#/definitions/model.PetHabit"
                },
                "room_care_habit": {
                    "$ref": "#/definitions/model.RoomCareHabit"
                },
                "sleep_habit": {
                    "$ref": "#/definitions/model.SleepHabit"
                },
                "smoke_habit": {
                    "$ref": "#/definitions/model.SmokeHabit"
                },
                "study_habit": {
                    "$ref": "#/definitions/model.StudyHabit"
                }
            }
        },
        "model.PetHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.PetPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.RoomCareHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.RoommatePreference": {
            "type": "object",
            "properties": {
                "other_preference": {
                    "type": "string"
                },
                "preferred_gender": {
                    "type": "string"
                },
                "preferred_pet_habit": {
                    "$ref": "#/definitions/model.PetHabit"
                },
                "preferred_room_care_habit": {
                    "$ref": "#/definitions/model.RoomCareHabit"
                },
                "preferred_sleep_habit": {
                    "$ref": "#/definitions/model.SleepHabit"
                },
                "preferred_smoke_habit": {
                    "$ref": "#/definitions/model.SmokeHabit"
                },
                "preferred_study_habit": {
                    "$ref": "#/definitions/model.StudyHabit"
                }
            }
        },
...
        "model.SleepHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.SmokeHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.SocialMedia": {
            "type": "object",
            "properties": {
                "facebook_url": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "twitter_url": {
                    "type": "string"
                }
            }
        },
        "model.Student": {
            "type": "object",
            "properties": {
                "biography": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enrollment_year": {
                    "type": "integer"
                },
                "faculty": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "personal_habit": {
                    "$ref": "#/definitions/model.PersonalHabit"
                },
                "pet_pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PetPicture"
                    }
                },
                "picture_url": {
                    "type": "string"
                },
                "roommate_preference": {
                    "$ref": "#/definitions/model.RoommatePreference"
                },
                "social_media": {
                    "$ref": "#/definitions/model.SocialMedia"
                }
            }
        },
        "model.StudyHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}